refactor bootstrap.jsx -> base.jsx
